# avoid buffering the output twice when writing to lepton file