--- a/Cargo.toml
+++ b/Cargo.toml
@@ -38,6 +38,7 @@ simple_logger ="5.0"
 unroll = "0.1"
 rayon-core = { version = "1", optional = true }
 git-version = "0.3"
+fixed-capacity-vec = "1.0.1"
 
 [target.'cfg(windows)'.dependencies]
 cpu-time = "1.0"