Track the carry overflow to we can immediately write out bytes to output stream and not have to worry about adjusting for carry.