@Melirius 2024-11-20
c5b693853c93972d48c423ab5b4bc49137c89352 helps, but still much slower than main
```
2024-11-20T20:56:19.974Z INFO  [lepton_jpeg::structs::lepton_file_writer] compressing to Lepton format
2024-11-20T20:56:20.453Z INFO  [lepton_jpeg::structs::lepton_file_writer] Number of threads: 8
2024-11-20T20:56:21.944Z INFO  [lepton_jpeg::structs::lepton_file_writer] worker threads 9985ms of CPU time in 1490ms of wall time
2024-11-20T20:56:21.944Z INFO  [lepton_jpeg::structs::lepton_file_writer] decompressing to verify contents
2024-11-20T20:56:23.567Z INFO  [lepton_jpeg_util] compressed input 22171278, output 17324076 bytes (compression = 28.0%)
2024-11-20T20:56:23.567Z INFO  [lepton_jpeg_util] Main thread CPU: 3593ms, Worker thread CPU: 21155 ms, walltime: 3593 ms

 Performance counter stats for 'taskset -c 10 nice -n -20 target/release/lepton_jpeg_util images/img_52MP_7k.jpg images/img_52MP_7k2.lep':

       836 854 716      cache-references                                                        (41,97%)
        75 722 326      cache-misses                     #    9,05% of all cache refs           (41,79%)
    16 120 935 933      cycles                                                                  (41,65%)
       956 121 210      ic_fetch_stall.ic_stall_back_pressure                                        (41,42%)
     1 106 567 900      stalled-cycles-frontend          #    6,86% frontend cycles idle        (41,53%)
    39 302 320 657      instructions                     #    2,44  insn per cycle            
                                                  #    0,03  stalled cycles per insn     (41,52%)
     4 328 449 995      branch-instructions                                                     (41,74%)
       166 308 719      branch-misses                    #    3,84% of all branches             (42,04%)
     5 614 369 841      ic_fetch_stall.ic_stall_any                                             (42,26%)
        42 227 531      ic_fetch_stall.ic_stall_dq_empty                                        (42,33%)
        62 449 857      l2_cache_misses_from_ic_miss                                            (42,46%)
     2 284 954 295      l2_latency.l2_cycles_waiting_on_fills                                        (42,25%)
           183 975      faults                                                                
                 1      migrations                                                            

       3,628224562 seconds time elapsed

       3,283399000 seconds user
       0,342937000 seconds sys
```